+++ custom_components/mopeka/device.py
@@ -2,11 +2,10 @@
 
 from __future__ import annotations
 
-from mopeka_iot_ble import DeviceKey
-
 from homeassistant.components.bluetooth.passive_update_processor import (
     PassiveBluetoothEntityKey,
 )
+from mopeka_iot_ble import DeviceKey
 
 
 def device_key_to_bluetooth_entity_key(